+++ src/evaluate.ts
@@ -523,7 +523,7 @@ async function loadDocument(
     void vscode.window.showTextDocument(doc, { preview: false });
   }
   const fileType = util.getFileType(doc);
-  const [ns, _] = namespace.getNamespace(doc, doc.positionAt(0));
+  const [ns, nsForm] = namespace.getNamespace(doc, doc.positionAt(0));
   const session = replSession.getSession(util.getFileType(doc));
 
   if (doc && doc.languageId == 'clojure' && fileType != 'edn' && getStateValue('connected')) {
@@ -531,7 +531,7 @@ async function loadDocument(
       ? await namespace.getUriForNamespace(session, ns)
       : doc.uri;
     const filePath = docUri.path;
-    return await loadFile(filePath, ns, pprintOptions, fileType);
+    return await loadFile(filePath, ns, nsForm, pprintOptions, fileType);
   }
 }
 
@@ -549,6 +549,7 @@ async function loadFileCommand() {
 async function loadFile(
   filePath: string,
   ns: string,
+  nsForm: string,
   pprintOptions: PrettyPrintingOptions,
   fileType: string
 ) {
@@ -611,7 +612,13 @@ async function loadFile(
     replSession.updateReplSessionType();
     if (getConfig().autoEvaluateCode.onFileLoaded[fileType]) {
       output.appendLineOtherOut(`Evaluating \`autoEvaluateCode.onFileLoaded.${fileType}\``);
-      const context = customSnippets.makeContext(vscode.window.activeTextEditor, ns, ns, fileType);
+      const context = customSnippets.makeContext(
+        vscode.window.activeTextEditor,
+        ns,
+        ns,
+        nsForm,
+        fileType
+      );
       await customSnippets.evaluateSnippet(
         util.getActiveTextEditor(),
         getConfig().autoEvaluateCode.onFileLoaded[fileType],
@@ -735,6 +742,33 @@ async function evaluateInOutputWindow(code: string, sessionType: string, ns: str
   }
 }
 
+async function evaluateInCurrentEditor(
+  editor: vscode.TextEditor,
+  code: string,
+  sessionType: string,
+  ns: string,
+  options
+) {
+  const document = editor?.document;
+  if (document) {
+    const evalPos = editor.selection.active;
+    try {
+      const session = replSession.getSession(sessionType);
+      return await evaluateCodeUpdatingUI(code, {
+        ...options,
+        filePath: document.fileName,
+        session,
+        ns,
+        nsForm: options.nsForm ?? `(in-ns '${ns})`,
+        line: evalPos.line,
+        column: evalPos.character,
+      });
+    } catch (e) {
+      output.appendLineOtherErr('Evaluation failed.');
+    }
+  }
+}
+
 export default {
   interruptAllEvaluations,
   loadDocument,
@@ -756,6 +790,7 @@ export default {
   toggleEvaluationSendCodeToOutputWindow,
   instrumentTopLevelForm,
   evaluateInOutputWindow,
+  evaluateInCurrentEditor,
   evaluateReplWindowForm,
   initInspectorDataProvider,
 };