+++ test-data/projects/shadow-w-backend/src/main/core.cljs
@@ -2,16 +2,16 @@
   (:require [reagent.core :as r]
             [reagent.dom :as rdom]))
 
-(defonce app-state (r/atom {:text "Hello world!"}))
+(defonce app-state (r/atom {:text "Hello"}))
 
-(defn hello-world []
+(defn hello-world [x]
   [:div
-   [:h1 (:text @app-state)]
+   [:h1 (:text @app-state) " " x "!"]
    [:h3 "Edit this and watch it change!"]])
 
 (defn ^:dev/after-load start []
   (js/console.log "start")
-  (rdom/render [hello-world]
+  (rdom/render [hello-world "world"]
                (. js/document (getElementById "app"))))
 
 (defn ^:export init []